rename the Table component
